+++ src-tauri/src/core/settings/blender_installation_location/mod.rs
@@ -1,4 +1,6 @@
 mod commands;
 mod impls;
+mod sweep;
 pub use commands::*;
 pub use impls::*;
+pub use sweep::*;